# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_dormitory-life_core_internal_server_request_models.CreatePhotoResponse:
    properties:
      fileName:
        type: string
      filePath:
        type: string
      size:
        format: int64
        type: integer
      url:
        type: string
    type: object
  github_com_dormitory-life_core_internal_server_request_models.CreateReviewResponse:
    properties:
      description:
        type: string
      dormitory_id:
        type: string
      owner_id:
        type: string
      photos:
        items:
          $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.CreatePhotoResponse'
        type: array
      review_id:
        type: string
      title:
        type: string
    type: object
  github_com_dormitory-life_core_internal_server_request_models.CreateSupportRequest:
    properties:
      description:
        type: string
      dormitoryId:
        type: string
      title:
        type: string
      userId:
        type: string
    type: object
  github_com_dormitory-life_core_internal_server_request_models.CreateSupportResponse:
    properties:
      message:
        type: string
    type: object
...
  github_com_dormitory-life_core_internal_server_request_models.ErrorResponse:
    properties:
      details:
        items:
          type: string
        type: array
      error:
        type: string
    type: object
...
  github_com_dormitory-life_core_internal_server_request_models.FileInfo:
    properties:
      last_modified:
        type: string
      name:
        type: string
      path:
        type: string
      size:
        type: integer
      url:
        type: string
    type: object
...
  github_com_dormitory-life_core_internal_server_request_models.GetDormitoryReviewsResponse:
    properties:
      reviews:
        items:
          $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.Review'
        type: array
    type: object
  github_com_dormitory-life_core_internal_server_request_models.Review:
    properties:
      created_at:
        type: string
      description:
        type: string
      dormitory_id:
        type: string
      owner_id:
        type: string
      review_id:
        type: string
      review_photos:
        items:
          $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.FileInfo'
        type: array
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: API Core сервиса Dormitory Life
  title: Dormitory Life Core API
  version: "1.0"
paths:
...
  /core/dormitories/{dormitory_id}/reviews:
    get:
      description: Получение отзывов общежития
      produces:
      - application/json
      responses:
        "200":
          description: Отзывы
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.GetDormitoryReviewsResponse'
        "400":
          description: Неверные данные / параметры запроса
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
      summary: Получение отзывов
      tags:
      - Reviews
    post:
      consumes:
      - multipart/form-data
      description: Создание отзыва на общежитие
      parameters:
      - description: ID общежития
        in: path
        name: dormitory_id
        required: true
        type: string
      - description: Заголовок отзыва
        in: formData
        name: title
        required: true
        type: string
      - description: Описание отзыва
        in: formData
        name: description
        required: true
        type: string
      - collectionFormat: multi
        description: Фотографии отзыва
        in: formData
        items:
          type: file
        name: photos
        required: true
        type: array
      produces:
      - application/json
      responses:
        "201":
          description: Отзыв создано
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.CreateReviewResponse'
        "400":
          description: Некорректные данные формы или отсутствуют обязательные поля
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "403":
          description: Нет прав на действие
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать отзыв
      tags:
      - Reviews
  /core/dormitories/{dormitory_id}/reviews/{event_id}:
    delete:
      description: Удаляет отзыв на общежитие
      parameters:
      - description: ID общежития
        in: path
        name: dormitory_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Неверные данные / параметры запроса
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "403":
          description: Нет прав на действие
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить отзыв
      tags:
      - Reviews
...
  /core/dormitories/support:
    post:
      consumes:
      - application/json
      description: Создает новое обращение и отправляет сообщение на почту
      parameters:
      - description: Данные обращения
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.CreateSupportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Обращение отправлено
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.CreateSupportResponse'
        "400":
          description: Неверные данные / параметры запроса
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
      summary: Создание обращения в поддержку
      tags:
      - core
...
schemes:
- http
- https
swagger: "2.0"